fix tests Signed-off-by: dbulahov <bulahovdenis@gmail.com>
