+++ mondrian/src/test/java/mondrian/test/ParentChildHierarchyTest.java
@@ -990,8 +990,8 @@ class ParentChildHierarchyTest {
             + " `store`.`store_state` as `Store State`,"
             + " `store`.`store_city` as `Store City`,"
             + " `store`.`store_name` as `Store Name`,"
-            + " `position`.`pay_type` as `Pay Type`,"
             + " `store`.`store_type` as `Store Type`,"
+            + " `position`.`pay_type` as `Pay Type`,"           
             + " `employee`.`management_role` as `Management Role`,"
             + " `employee`.`position_title` as `Position Title`,"
             + " `department`.`department_id` as `Department Description`,"
@@ -1022,9 +1022,9 @@ class ParentChildHierarchyTest {
                 + " `Store Country` ASC,"
                 + " `Store State` ASC,"
                 + " `Store City` ASC,"
-                + " `Store Name` ASC,"
-                + " `Pay Type` ASC,"
+                + " `Store Name` ASC,"                
                 + " `Store Type` ASC,"
+                + " `Pay Type` ASC,"
                 + " `Management Role` ASC,"
                 + " `Position Title` ASC,"
                 + " `Department Description` ASC,"
